+++ app/src/main/AndroidManifest.xml
@@ -13,18 +13,17 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
 
-        <activity android:name=".AllEvents"></activity>
+        <activity android:name=".Event.AllEvents"></activity>
         <activity android:name=".ForgetPassword.ForgetPassword"></activity>
         <activity android:name=".MainActivity"></activity>
         <activity android:name=".Login.LoginActivity"></activity>
         <activity android:name=".Splash.SplashActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".MainActivity"></activity>
+
     </application>
 
 </manifest>
