+++ app/src/main/java/com/example/shiraz_uni_app/Event/Example.java
@@ -1,4 +1,4 @@
-package com.example.shiraz_uni_app;
+package com.example.shiraz_uni_app.Event;
 
 public class Example {
