+++ app/src/main/java/com/example/shiraz_uni_app/Event/ExampleAdapter.java
@@ -1,4 +1,4 @@
-package com.example.shiraz_uni_app;
+package com.example.shiraz_uni_app.Event;
 
 import android.support.annotation.NonNull;
 import android.support.v7.widget.RecyclerView;
@@ -6,6 +6,9 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+
+import com.example.shiraz_uni_app.R;
+
 import java.util.ArrayList;
 
 public class ExampleAdapter extends RecyclerView.Adapter<ExampleAdapter.ExampleViewHolder> {
