+++ app/src/main/res/drawable/date_background.xml
@@ -3,5 +3,5 @@
     android:shape="rectangle">
     <corners android:bottomLeftRadius="30dp"
         android:bottomRightRadius="30dp"/>
-    <solid android:color="@color/darkBlue" />
+    <solid android:color="@color/dark_blue" />
 </shape>
