+++ app/src/main/res/drawable/event_background.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <ripple xmlns:android="http://schemas.android.com/apk/res/android"
-    android:color="@color/darkGrey">
+    android:color="@color/dark_grey">
     <item android:id="@android:id/mask">
         <shape android:shape="rectangle">
             <solid android:color="@color/colorPrimaryDark" />
@@ -10,7 +10,7 @@
     <item android:id="@android:id/background" >
         <shape android:shape="rectangle" >
             <solid android:color="@color/white"/>
-            <stroke android:color="@color/darkBlue" android:width="1dp"/>
+            <stroke android:color="@color/dark_blue" android:width="1dp"/>
             <corners android:radius="30dp" />
         </shape>
     </item>
