+++ app/src/main/res/layout/activity_all_events.xml
@@ -5,12 +5,12 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".AllEvents">
+    tools:context=".Event.AllEvents">
 
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:background="@color/darkBlue"
+        android:background="@color/dark_blue"
         android:orientation="horizontal"
         android:paddingBottom="16dp"
         android:paddingTop="16dp"
