+++ app/src/main/res/layout/activity_event_item.xml
@@ -2,7 +2,6 @@
 <android.support.constraint.ConstraintLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="140dp"
     android:layout_marginLeft="20dp"
@@ -26,7 +25,7 @@
         android:maxLines="2"
         android:text="this is line 1 \n this is line 2"
         android:textAlignment="center"
-        android:textColor="@color/darkBlue"
+        android:textColor="@color/dark_blue"
         android:textSize="16sp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
