+++ app/src/main/res/layout/activity_forget_password.xml
@@ -119,9 +119,9 @@
         <RelativeLayout
             android:id="@+id/relativeLayout"
             android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:gravity="center_horizontal"
-            android:background="@drawable/back_button">
+            android:layout_height="38dp"
+            android:background="@drawable/back_button"
+            android:gravity="center_horizontal">
 
 
             <ImageButton
